feat(meta): allow build for both TS and JS
We use Babel to transpile the TS files to ensure we can
leverage the plugins (like for Ramda) and have the same
JavaScript target for both JS and TS files.

Because we copy the Flow files next to the JS built files
in lib, we don't need to include them from the `src`
directory.