feat: convert records/config to TS
NOTE: some tests were fixed in the process, some
unexpected behaviours were because the objects
passed to the function were not formed correctly.

This removed a few TODOs.

I still left a few of them with `@ts-expect-error`
to test for improper arguments. Not sure of their
value, but because we rely on Ramda so much, it's
probably a good thing to know what happens in these
edge cases (especially when we use `@types/ramda`
which could potentially be out of sync/incomplete
with the version of Ramda we're using).